+++ example/lib/main.dart
@@ -1,10 +1,11 @@
 import 'dart:developer';
 
+import 'package:example/colors.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:simple_sidebar/simple_sidebar.dart';
 import 'package:simple_sidebar/simple_sidebar_item.dart';
-import 'package:simple_sidebar/simple_sidebar_theme.dart';
+import 'package:simple_sidebar/simple_sidebar_theme_data.dart';
 
 void main() {
   runApp(const MyApp());
@@ -19,43 +20,32 @@ class MyApp extends StatelessWidget {
       title: 'Simplesidebar Demo',
       theme: ThemeData(
         textTheme: const TextTheme(
-          bodyText1: TextStyle(color: Colors.white),
-          bodyText2: TextStyle(color: Colors.white),
+          bodyLarge: TextStyle(color: Colors.white),
+          bodyMedium: TextStyle(color: Colors.white),
         ),
       ),
       debugShowCheckedModeBanner: kDebugMode,
-      home: MyHomePage(title: 'Flutter Demo Home Page'),
+      home: MyHomePage(),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  final List<SimpleSidebarItem> sidebarItems = [
-    SimpleSidebarItem(
-        title: "Home",
-        iconFront: Icons.home_outlined,
-        child: const Center(child: Text("Home"))),
-    SimpleSidebarItem(
-        title: "Gallery",
-        iconFront: Icons.image_outlined,
-        child: const Center(child: Text("Gallery"))),
-    SimpleSidebarItem(
-        title: "Users",
-        iconFront: Icons.group_outlined,
-        child: const Center(child: Text("Users"))),
-    SimpleSidebarItem(
-        title: "Extra long name for a menu item",
-        wrapWord: true,
-        iconFront: Icons.group_outlined,
-        child: const Center(child: Text("Users"))),
-    SimpleSidebarItem(
-        title: "Exit",
-        iconEnd: Icons.close,
-        child: const Center(child: Text("Exit"))),
-  ];
-  final String title;
+  MyHomePage({super.key});
 
-  MyHomePage({super.key, required this.title});
+  final SimpleSidebarThemeData theme = SimpleSidebarThemeData()
+    ..expandedShape = const BorderRadius.all(Radius.circular(8))
+    ..collapsedShape = const BorderRadius.all(Radius.circular(8));
+  final List<SimpleSidebarItem> sidebarItems = [
+    SimpleSidebarItem(onTap: () {}, title: "Home", leading: Icons.home_outlined),
+    SimpleSidebarItem(onTap: () {}, title: "Gallery", leading: Icons.image_outlined),
+    SimpleSidebarItem(onTap: () {}, title: "Users", leading: Icons.group_outlined),
+    SimpleSidebarItem(onTap: () {}, title: "Extra long name for a menu item", leading: Icons.list),
+  ];
+  final List<SimpleSidebarItem> footerItems = [
+    SimpleSidebarItem(onTap: () {}, title: "Settings", leading: Icons.settings),
+    SimpleSidebarItem(onTap: () {}, title: "Logout", leading: Icons.close),
+  ];
 
   @override
   State<MyHomePage> createState() => _MyHomePageState();
@@ -67,43 +57,55 @@ class _MyHomePageState extends State<MyHomePage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        backgroundColor: Colors.blueGrey[800],
-        body: Row(
-          children: [
-            SimpleSidebar(
-              simpleSidebarTheme: SimpleSidebarTheme(),
-              initialExpanded: false,
-              sidebarItems: widget.sidebarItems,
-              onTapped: (value) => onTapped(value),
-              toggleSidebar: (value) {
-                log("Sidebar is now $value");
-              },
+      backgroundColor: kcBackground,
+      body: Row(
+        children: [
+          SimpleSidebar(
+            theme: widget.theme,
+            initialExpanded: true,
+            initialIndex: 1,
+            childrens: widget.sidebarItems,
+            footerItems: widget.footerItems,
+            onTap: (value) {
+              setState(() {
+                selected = value;
+              });
+              log("selected value $value");
+            },
+          ),
+          Expanded(
+            child: AnimatedOpacity(
+              opacity: isVisible ? 1 : 0,
+              duration: const Duration(milliseconds: 300),
+              child: Container(
+                  margin: const EdgeInsets.all(8),
+                  child: const <Widget>[
+                    Text("hallo"),
+                    Text("hallo2"),
+                    Text("hallo3"),
+                    Text("hallo4"),
+                    Text("hallo5"),
+                    Text("hallo6"),
+                  ][selected]),
             ),
-            Expanded(
-              child: AnimatedOpacity(
-                opacity: isVisible ? 1 : 0,
-                duration: const Duration(milliseconds: 400),
-                child: Container(
-                    margin: const EdgeInsets.all(8),
-                    child: widget.sidebarItems.elementAt(selected).child),
-              ),
-            )
-          ],
-        ));
+          ),
+        ],
+      ),
+    );
   }
 
-  void onTapped(int value) {
-    setState(() {
-      isVisible = false;
-    });
-    Future.delayed(const Duration(milliseconds: 400), () {
-      setState(() {
-        selected = value;
-      });
-    }).then((value) {
-      setState(() {
-        isVisible = true;
-      });
-    });
-  }
+  // void onTapped(int value) {
+  //   setState(() {
+  //     isVisible = false;
+  //   });
+  //   Future.delayed(const Duration(milliseconds: 400), () {
+  //     setState(() {
+  //       selected = value;
+  //     });
+  //   }).then((value) {
+  //     setState(() {
+  //       isVisible = true;
+  //     });
+  //   });
+  // }
 }