+++ lib/simple_sidebar_element.dart
@@ -1,4 +1,0 @@
-/// Interface like base class
-abstract class SimpleSidebarElement {
-  SimpleSidebarElement();
-}